+++ internal/database/stocks.go
@@ -9,7 +9,7 @@ import (
 )
 
 // CreateStock creates a new stock with the given name and initial price and returns the ID of the new stock. The newly created stock is active.
-func CreateStock(name string, initPrice float64) (int64, error) {
+func CreateStock(name string, initPrice int64) (int64, error) {
 	currentTimeStamp := time.Now()
 
 	db := getDB()
